tests/wpt/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-single-value-in-body.html

<!doctype html>
<title>Meta color-scheme in body should apply</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/semantics.html#meta-color-scheme">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/compute-root-color-scheme.js"></script>
<body>
  <meta name="color-scheme" content="dark">
</body>
<script>
  assert_root_color_scheme("dark", "Meta color-scheme in body should apply.");
</script>
